+++ server/impl/src/main/java/com/walmartlabs/concord/server/cfg/SecretStoreConfiguration.java
@@ -9,9 +9,9 @@ package com.walmartlabs.concord.server.cfg;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -47,6 +47,10 @@ public class SecretStoreConfiguration implements Serializable {
     @Config("secretStore.maxEncryptedStringLength")
     private int maxEncryptedStringLength;
 
+    @Inject
+    @Config("secretStore.keySize")
+    private int keySize;
+
     public byte[] getServerPwd() {
         return serverPwd;
     }
@@ -61,5 +65,9 @@ public class SecretStoreConfiguration implements Serializable {
 
     public int getMaxEncryptedStringLength() {
         return maxEncryptedStringLength;
+    }
+
+    public int getKeySize() {
+        return keySize;
     }
 }