+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- concord-server: configurable key size for generated key pairs;
 - concord-server: expose Jetty Sessions metrics;
 - concord-task: support for `outVars` for `action: fork` when
 `suspend` is enabled;